#refactor add utils.h file to gather util functions
