+++ src/core/service.h
@@ -33,58 +33,6 @@
 #include "session/session.h"
 #include "session/udpforwardsession.h"
 
-#ifdef ENABLE_REUSE_PORT
-typedef boost::asio::detail::socket_option::boolean<SOL_SOCKET, SO_REUSEPORT> reuse_port;
-#endif // ENABLE_REUSE_PORT
-
-#ifndef IP_RECVTTL
-#define IP_RECVTTL 12
-#endif
-
-#ifndef IPV6_RECVHOPLIMIT
-#define IPV6_RECVHOPLIMIT 51
-#endif
-
-#ifndef IPV6_HOPLIMIT
-#define IPV6_HOPLIMIT 21
-#endif
-
-#ifndef IP_TTL
-#define IP_TTL 4
-#endif
-
-// copied from shadowsocks-libe udprelay.h
-#ifndef IP_TRANSPARENT
-#define IP_TRANSPARENT       19
-#endif
-
-#ifndef IP_RECVORIGDSTADDR
-#ifdef  IP_ORIGDSTADDR
-#define IP_RECVORIGDSTADDR   IP_ORIGDSTADDR
-#else
-#define IP_RECVORIGDSTADDR   20
-#endif
-#endif
-
-#ifndef IPV6_RECVORIGDSTADDR
-#ifdef  IPV6_ORIGDSTADDR
-#define IPV6_RECVORIGDSTADDR   IPV6_ORIGDSTADDR
-#else
-#define IPV6_RECVORIGDSTADDR   74
-#endif
-#endif
-
-#ifndef SOL_IP
-#define SOL_IP  IPPROTO_IP
-#endif
-
-#ifndef SOL_IPV6
-#define SOL_IPV6  IPPROTO_IPV6
-#endif
-
-#define PACKET_HEADER_SIZE (1 + 28 + 2 + 64)
-#define DEFAULT_PACKET_SIZE 1397 // 1492 - PACKET_HEADER_SIZE = 1397, the default MTU for UDP relay
-
 class Pipeline;
 class icmpd;
 class Service {
@@ -125,116 +73,4 @@ public:
 
     Pipeline* search_default_pipeline();
 };
-
-
-template<typename ThisT, typename EndPoint>
-void connect_out_socket(ThisT this_ptr, std::string addr, std::string port, boost::asio::ip::tcp::resolver& resolver, 
-    boost::asio::ip::tcp::socket& out_socket, EndPoint in_endpoint, std::function<void()> connected_handler){
-    
-    resolver.async_resolve(addr, port, [=, &out_socket]
-     (const boost::system::error_code error, boost::asio::ip::tcp::resolver::results_type results) {
-        if (error || results.empty()) {
-            _log_with_endpoint(in_endpoint, "cannot resolve remote server hostname " + addr + ":" + port + " reason: " + error.message(), Log::ERROR);
-            this_ptr->destroy();
-            return;
-        }
-        auto iterator = results.begin();
-        _log_with_endpoint(in_endpoint, addr + " is resolved to " + iterator->endpoint().address().to_string(), Log::ALL);
-        boost::system::error_code ec;
-        out_socket.open(iterator->endpoint().protocol(), ec);
-        if (ec) {
-            this_ptr->destroy();
-            return;
-        }
-        if (this_ptr->config.tcp.no_delay) {
-            out_socket.set_option(boost::asio::ip::tcp::no_delay(true));
-        }
-        if (this_ptr->config.tcp.keep_alive) {
-            out_socket.set_option(boost::asio::socket_base::keep_alive(true));
-        }
-#ifdef TCP_FASTOPEN_CONNECT
-        if (this_ptr->config.tcp.fast_open) {
-            using fastopen_connect = boost::asio::detail::socket_option::boolean<IPPROTO_TCP, TCP_FASTOPEN_CONNECT>;
-            boost::system::error_code ec;
-            out_socket.set_option(fastopen_connect(true), ec);
-        }
-#endif // TCP_FASTOPEN_CONNECT
-        auto timeout_timer = std::shared_ptr<boost::asio::steady_timer>(nullptr);
-        if(this_ptr->config.tcp.connect_time_out > 0){
-            // out_socket.async_connect will be stuck forever when the host is not reachable
-            // we must set a timeout timer
-            timeout_timer = std::make_shared<boost::asio::steady_timer>(out_socket.get_io_context());
-            timeout_timer->expires_after(std::chrono::seconds(this_ptr->config.tcp.connect_time_out));
-            timeout_timer->async_wait([=](const boost::system::error_code error) {
-                if(!error){
-                    _log_with_endpoint(in_endpoint, "cannot establish connection to remote server " + addr + ':' + port + " reason: timeout", Log::ERROR);
-                    this_ptr->destroy();
-                }
-            });
-        }
-
-        out_socket.async_connect(*iterator, [=](const boost::system::error_code error) {
-            if(timeout_timer){
-                timeout_timer->cancel();
-            }
-
-            if (error) {
-                _log_with_endpoint(in_endpoint, "cannot establish connection to remote server " + addr + ':' + port + " reason: " + error.message(), Log::ERROR);
-                this_ptr->destroy();
-                return;
-            }        
-
-            connected_handler();
-        });        
-    });
-}
-
-template<typename ThisT, typename EndPoint>
-void connect_remote_server_ssl(ThisT this_ptr, std::string addr, std::string port, boost::asio::ip::tcp::resolver& resolver, 
-    boost::asio::ssl::stream<boost::asio::ip::tcp::socket>& out_socket, EndPoint in_endpoint, std::function<void()> connected_handler){
-    
-    connect_out_socket(this_ptr, addr, port, resolver, out_socket.next_layer(), in_endpoint, [=, &out_socket](){
-        out_socket.async_handshake(boost::asio::ssl::stream_base::client, [=, &out_socket](const boost::system::error_code error) {
-            if (error) {
-                _log_with_endpoint(in_endpoint, "SSL handshake failed with " + addr + ':' + port + " reason: " + error.message(), Log::ERROR);
-                this_ptr->destroy();
-                return;
-            }
-            _log_with_endpoint(in_endpoint, "tunnel established");
-            if (this_ptr->config.ssl.reuse_session) {
-                auto ssl = out_socket.native_handle();
-                if (!SSL_session_reused(ssl)) {
-                    _log_with_endpoint(in_endpoint, "SSL session not reused");
-                } else {
-                    _log_with_endpoint(in_endpoint, "SSL session reused");
-                }
-            }
-            connected_handler();
-        });        
-    });
-}
-
-template<typename ThisPtr>
-void shutdown_ssl_socket(ThisPtr this_ptr, boost::asio::ssl::stream<boost::asio::ip::tcp::socket>& socket){
-    if (socket.next_layer().is_open()) {
-        auto self = this_ptr->shared_from_this();
-        auto ssl_shutdown_timer = std::make_shared<boost::asio::steady_timer>(socket.next_layer().get_io_context());
-        auto ssl_shutdown_cb = [self, ssl_shutdown_timer, &socket](const boost::system::error_code error) {
-            if (error == boost::asio::error::operation_aborted) {
-                return;
-            }
-            boost::system::error_code ec;
-            ssl_shutdown_timer.get()->cancel();
-            socket.next_layer().cancel(ec);
-            socket.next_layer().shutdown(boost::asio::ip::tcp::socket::shutdown_both, ec);
-            socket.next_layer().close(ec);
-        };
-        boost::system::error_code ec;
-        socket.next_layer().cancel(ec);
-        socket.async_shutdown(ssl_shutdown_cb);
-        ssl_shutdown_timer.get()->expires_after(std::chrono::seconds(30));
-        ssl_shutdown_timer.get()->async_wait(ssl_shutdown_cb);
-    }    
-}
-
 #endif // _SERVICE_H_
